ARM: tegra: ventana: disable usb1 vbus gpio
disable the usb1 vbus gpio in kernel and not depend on bootloader
to disable the vbus initially.
Currently, the kernel enables and disables the vbus when
the otg is configured to host mode, but when the system boots
with no cable connected the kernel relies on the bootloader
to turn the vbus off.
This CL removes that dependency.

Bug 1047048
Signed-off-by: Preetham Chandru R <pchandru@nvidia.com>
Reviewed-on: http://git-master/r/#change,141970
(cherry picked from commit c3461995dd156968d766ec05879fd1097221ceb8)
Change-Id: I96c1f4b97a2cafdfd498b591647200d26298a43b
Reviewed-on: http://git-master/r/142839
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Kiran Adduri <kadduri@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Matthew Pedro <mapedro@nvidia.com>